requirements.txt
Flask==0.12
gevent==1.2.1
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.5
locustio==0.8.0
MarkupSafe==1.0
msgpack-python==0.4.8
pyquery==1.2.17
pyzmq==16.0.2
requests==2.13.0
Werkzeug==0.11.15
